fall_model_service/requirements.txt
flask>=3.0.3
google-cloud-storage>=2.16.0
opencv-python-headless>=4.10.0
torch>=2.2.0
torchvision>=0.17.0
